groceries no longer inherit from Food, created separate List Item for Groceries

+++ app/src/main/java/com/kunle/aisle9b/models/Meal.kt
@@ -12,7 +12,7 @@ data class Meal(
     val servingSize: String,
     val mealPic: Uri? = null,
     val notes: String,
-    val apiID: Long = -1,
+    val apiID: Int = -1,
     val visible: Boolean = true
 ) {
     constructor(name: String, servingSize: String, apiID: Int) :
